+++ confluence-client/src/main/kotlin/com/github/zeldigas/confclient/ConfluenceClientConfig.kt
@@ -1,5 +1,6 @@
 package com.github.zeldigas.confclient
 
+import io.ktor.client.plugins.logging.*
 import io.ktor.http.*
 
 /**
@@ -8,5 +9,7 @@ import io.ktor.http.*
 data class ConfluenceClientConfig(
     val server: Url,
     val skipSsl: Boolean,
-    val auth: ConfluenceAuth
+    val auth: ConfluenceAuth,
+    val httpLogLevel: LogLevel = LogLevel.NONE,
+    val requestTimeout: Long? = null,
 )